rewards/robo_instruct/train/train_requirements.txt
transformers==4.49.0
accelerate==1.4.0
numpy==1.24.4
hydra-core==1.3.2
trl==0.15.2
wandb==0.19.8
unsloth==2025.3.8
unsloth_zoo==2025.3.7